requirements.txt
transformers
transformers[sentencepiece]
dataset
sacrebleu
rouge_score 
py7zr
pandas 
nltk 
tqdm
PyYAML
torch
notebook
boto3 
ensure
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.1
fastapi==0.78.0
uvicorn==0.18.3
jinja2==3.1.2

-e . 